+++ grid2op/Exceptions/AttentionBudgetExceptions.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2021, RTE (https://www.rte-france.com)
+# Copyright (c) 2019-2020, RTE (https://www.rte-france.com)
 # See AUTHORS.txt
 # This Source Code Form is subject to the terms of the Mozilla Public License, version 2.0.
 # If a copy of the Mozilla Public License, version 2.0 was not distributed with this file,
@@ -6,6 +6,12 @@
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
 
-__all__ = ["LinearAttentionBudget"]
+from grid2op.Exceptions.Grid2OpException import Grid2OpException
 
-from grid2op.operator_attention.attention_budget import LinearAttentionBudget
+
+class NotEnoughAttentionBudget(Grid2OpException):
+    """
+    This exception is raised when the player attempted to alert the "human operator" but it has not
+    enough budget to do so.
+    """
+    pass