+++ CHANGELOG.rst
@@ -25,7 +25,7 @@ Change Log
 
 [1.6.0] - 2021-xx-yy
 --------------------
-- [ADDED] support for the "alarm" feature
+- [ADDED] support for the "alarm operator" / "attention budget" feature
 
 [1.5.2] - 2021-xx-yy
 -----------------------